+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+v2.5.11
+-------
+* Added a variable decrement method.
+* Avoid generating a conversion instruction when comparing against a constant primitive value.
+
 v2.5.10 (2024-08-07)
 -------
 * Added a convenience method for initializing static fields to complex constants.